Clean up the dataset module
- Split up in submodules.
- Remove unnecessary use of generics.
- Other simplifications.
